+++ lib/solid_queue_guard/metrics/emit_depth_job.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module SolidQueueGuard
+  module Metrics
+    class EmitDepthJob < ActiveJob::Base
+      queue_as :solid_queue_recurring
+
+      def perform
+        DepthEmitter.emit!
+      end
+    end
+  end
+end